Refactor structure for contact us page
