+++ src/pages/contactUs/components/ContactUsContent.js
@@ -1,6 +1,5 @@
 import React from 'react';
-import styled, { css } from 'styled-components';
-import media from '@kiwicom/orbit-components/lib/utils/mediaQuery';
+import styled from 'styled-components';
 import { Heading, Text, TextLink } from '@kiwicom/orbit-components/lib';
 import { contactUsGoogleFormPath } from '@constants/googleFormPaths';
 import { MaxWidthContainer } from '@components/containers';
