+++ src/pages/contactUs/index.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import ContactUsContent from '../modules/ContactUsContent';
+import ContactUsContent from './components/ContactUsContent';
 
 const ContactUsPage = () => {
   return <ContactUsContent />;
